feat(): link dsp to gui (#13)
* refactor(dsp): informative filenames

* feat(piano): gui piano is now partially linked with dsp backend

* feat(adsr): partially linked gui adsr with dsp backend

* void(): adsr improved, bugfixes, linked volume slider & audio button with backend, visualisation for piano, ...

* fix: disable playbackMode button

* chore(): format files

---------

+++ dsp/dsp_utils.c
@@ -1,9 +1,9 @@
-#include "utils_dsp.h"
+#include "dsp_utils.h"
 #include <math.h>
 
 // midi to freq
 float midi_to_freq(int midiNote) {
-  return 440.0f * powf(2.0f, (midiNote - 69) / 12.0f);
+  return 440.0f * powf(2.0f, ((float)midiNote - 69.0f) / 12.0f);
 }
 
 // prevent clipping
